mtk_parser.py
import requests
from bs4 import BeautifulSoup
import logging

logger = logging.getLogger(__name__)

def parse_schedule_from_mtk(class_name):
    """Парсинг расписания с сайта МТК"""
    try:
        # Здесь должен быть твой код парсинга
        # Пример:
        url = f"http://site.com/schedule/{class_name}"
        response = requests.get(url)
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Парсим данные
        schedule = {}
        # ... твоя логика ...
        
        return schedule
    except Exception as e:
        logger.error(f"Ошибка парсинга: {e}")
        return None